Better theme file finding algorithm for ST3
In Sublime Text 3, Themr now makes use of the new API method `sublime.find_resources()` to find all loaded resources of the pattern "*.sublime-theme". This is more reliable than the old directory walking method because it always finds all .sublime-theme files ST3 has loaded, and only the .sublime-theme files ST3 has loaded.

In ST2, sublime.find_resources() is not present, so at runtime we check the version of ST and fall back to the old method if < 3.0.

Additionally, because Themr now always finds all loaded packages, we never have to manually add the default theme to the list (which was required before in ST3 since the default theme existed in Sublime's internal 'Default' package, rather than in the user's packages directory.)